requirements.txt
openai-whisper>=20240930
openai>=1.40.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
pyzmq>=25.0
pytest>=7.0
